functional import for scenarios, and a merge to combine objects
