+++ index.html
@@ -16,5 +16,6 @@
             <span id="U">U</span>
             <span id="S">S</span>
         </div>
-	</body>
+    </body>
+    <script src="./scenario_controller/scenario_controller.js"></script>
 </html>
