+++ package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "description": "",
   "main": "server.js",
+  
   "scripts": {
     "start": "node server.js"
   },
@@ -18,4 +19,5 @@
     "nodemon": "^2.0.4",
     "sequelize": "^4.44.4"
   }
+  
 }
