+++ public/assets/rpg.html
@@ -45,6 +45,7 @@
                 </div>
             </div>
         </div> 
-        <script src="./js/scenario_controller.js"></script>
+        <script type="module" src="./scenario_controller/scenario_controller.js"></script>
+
 	</body>
 </html>
