+++ public/assets/scenario_controller/index.html
@@ -1,0 +1,4 @@
+<script src="./scenario_controller.js"></script>
+<script src="./scenarios/1_starting_scenario.js"></script>
+<script src="./scenarios/2_scenario_hallway2.js"></script>
+<script src="./scenarios/3_operation_scenario.js"></script>
