+++ public/assets/scenario_controller/scenarios/2_scenario_hallway2.js
@@ -36,3 +36,4 @@ let scenario_2={
             killPlayer: false
         }]
 };
+export default {scenario_2};
